add dagger2x

+++ app/src/main/java/everis/rx/com/rxandroid/fr/MultiplicarFragment.java
@@ -63,7 +63,7 @@ public class MultiplicarFragment extends BaseFragment {
     public void onStartChangePasswordClick() {
         int ant = Integer.parseInt(a.getText().toString());
         int bnt = Integer.parseInt(b.getText().toString());
-        c.setText(""+(ant*bnt));
+        c.setText(""+Integer.valueOf(ant*bnt));
 
     }
